+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improve `filesystem` server for better compatibility with the official implementation.
 - Improve error message in `JSONRPCError` struct.
 - Set message in `JSONRPCError` struct to string representation instead of generic "Internal error", except for `CallTool` request which returns `CallToolResult` with `IsError` set to true and the error setted as the content.
+- Skip `.git` directory in `directory_tree` call tool request at `filesystem` server.
 
 ## [0.5.0] - 2025-02-09